+++ src/test/scala/edu/ie3/simona/service/weather/WeatherSourceWrapperSpec.scala
@@ -28,9 +28,9 @@ import edu.ie3.simona.service.weather.WeatherSourceWrapperSpec.*
 import edu.ie3.simona.test.common.UnitSpec
 import edu.ie3.util.geo.GeoUtils
 import edu.ie3.util.interval.ClosedInterval
-import edu.ie3.util.scala.quantities.{Irradiance, WattsPerSquareMeter}
 import org.locationtech.jts.geom.Point
 import squants.motion.MetersPerSecond
+import squants.radio.{Irradiance, WattsPerSquareMeter}
 import squants.thermal.{Celsius, Kelvin}
 import squants.{Temperature, Velocity}
 import tech.units.indriya.ComparableQuantity